Added teams and tasks to roster population.
Also broke tasks down further for profile population. Tasks are now
split between accepted and unaccepted for both ownerTasks and
workerTasks. The new fields are as follows:
ownerTasksUnaccepted
ownerTasksAccepted
workerTasksUnaccepted
workerTasksAccepted

If needed, I can also try to add tasks that the user can accept because
they are on the team the task is assigned to.