Updated task owners/workers slightly
Task owners and workers which are users now have a status field instead
of an accepted field. Status can be one of the following:
'Accepted','Pending'(default),'Rejected'. When attaching tasks for
reading in the user's profile, the tasks are now just attached to the
same variables they were before: ownerTasks and workerTasks. They now
come split into two parts: task (the task object) and status(the user's
assignment status for that task). You will need to dump tasks to use.